Refactor code into reusable package
