+++ README.md
@@ -4,10 +4,11 @@ This repository provides a small set of scripts for monitoring window focus acti
 
 ## Contents
 
-* `standalone_focus_monitor.py` – a command line tool that records the currently focused window over time. It writes log files and daily summaries to a `focus_logs` directory.
+* `standalone_focus_monitor.py` – a command line entry point that runs the monitor logic from the `focus_tracker` package.
   The monitor can optionally grab screenshots and OCR text for selected applications and organises activity into five‑minute buckets with automatic summaries. Daily summaries now include a **focus score**, detected meeting time and per‑app classifications (productive vs distraction). Browser profiles in Chrome and Edge are detected so work and personal usage can be separated.
 * `focus_monitor_dashboard.py` – a [Streamlit](https://streamlit.io/) dashboard that reads the log files generated by the tool and displays interactive charts.
   The dashboard includes a label editor for renaming activities, a page for reviewing five‑minute summaries with personal notes, and new tools for managing activity categories.  Categories can be suggested automatically using a local LLM and applied to historical data.  An optional test page helps diagnose category suggestions.
+* `focus_tracker/` – a Python package containing the monitoring agent and dashboard utilities.
 
 ## Requirements
